requirements.txt
gspread==5.7.2
gspread_dataframe==3.3.0
mysql-connector-python==8.0.32
pandas==1.5.3
protobuf==4.22.0
PyDrive==1.3.1
